test(api): isolate unix socket tests behind linux build tag
